refactor(desktop): decompose monolithic store into domain slices
Extract 7 self-contained action groups from store.ts into dedicated
slice files under src/store/slices/: notifications, plans, budget,
skills, diff-comments, github, sidebar. Each slice exports a
createXxxSlice factory. AppState/AppActions interfaces unchanged.
